reorganize internal modules
