+++ src/pms/sensor/__init__.py
@@ -1,2 +1,8 @@
-from .sensor import Sensor  # isort: skip
-from .reader import MessageReader, SensorReader
+from pms import logger
+from pms.core import MessageReader, Sensor, SensorReader
+
+logger.warning(
+    f"the pms.sensor module is deprecated, import from pms.core instead",
+    DeprecationWarning,
+    2,
+)
