+++ src/pms/sensors/bosch_sensortec/mcu680.py
@@ -8,7 +8,7 @@ from dataclasses import dataclass, field
 from typing import Tuple
 
 from pms import SensorWarmingUp, WrongMessageChecksum, WrongMessageFormat
-from pms.sensor import base
+from pms.core import base
 
 ALIASES = ("BME680",)
