+++ src/pms/sensors/novafitness/extra_commands.py
@@ -5,7 +5,7 @@ https://learn.watterott.com/sensors/sds011/sds011_protocol.pdf
 """
 
 
-from pms.sensor.base import Cmd
+from pms.core.base import Cmd
 
 
 def _msg(cmd: int, payload: str, device: int = 0xFFFF) -> bytes:
