+++ src/pms/sensors/novafitness/sds01x.py
@@ -8,7 +8,7 @@ from dataclasses import dataclass, field
 from typing import Tuple
 
 from pms import SensorWarmingUp, WrongMessageChecksum, WrongMessageFormat
-from pms.sensor import base
+from pms.core import base
 
 ALIASES = ("SDS011", "SDS018", "SDS021")
