+++ src/pms/sensors/novafitness/sds198.py
@@ -6,7 +6,7 @@ NovaFitness SDS198 sensors
 
 from dataclasses import dataclass, field
 
-from pms.sensor import base
+from pms.core import base
 
 from . import sds01x
