+++ src/pms/sensors/senserion/sps30.py
@@ -11,7 +11,7 @@ from dataclasses import dataclass, field
 from typing import Tuple
 
 from pms import SensorWarmingUp, WrongMessageChecksum, WrongMessageFormat
-from pms.sensor import base
+from pms.core import base
 
 BAUD = 115_200
